requirements.txt
fastapi
sqlalchemy
aiosqlite
uvicorn
pydantic
pydantic[email]
jinja2
pydantic-settings
alembic
bcrypt
passlib[bcrypt]
python-jose
loguru